move packages, fix checkstyle, add README
